WIP generating locations and saving them to the memory db
For the time being the "memory db" is actually written to db to check
the results. This creates the `memory.db' file in the current
directory. Make sure to delete the database between runs, or the
program will quit with a segfault and/or a dumped core and/or some
other nasty thing...